backend/AgileHackathonBackend/src/main/java/com/agilehackathon/status/StatusServlet.java
package com.agilehackathon.status;

import javax.servlet.*;
import javax.servlet.http.HttpServlet;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;

public class StatusServlet extends HttpServlet {

    public static final String SUCCESS_MESSAGE = "OK DUDE";

    @Override
    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
        response.setStatus(HttpServletResponse.SC_OK);
        response.getWriter().print(SUCCESS_MESSAGE);
    }
}
