backend/AgileHackathonBackend/src/test/java/com/agilehackathon/status/StatusServletTest.java
package com.agilehackathon.status;

import org.junit.Test;

import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;

import java.io.PrintWriter;

import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;

public class StatusServletTest {
    HttpServletRequest request = mock(HttpServletRequest.class);
    HttpServletResponse response = mock(HttpServletResponse.class);
    PrintWriter printWriter = mock(PrintWriter.class);
    private final StatusServlet statusServlet = new StatusServlet();

    @Test
    public void statusPageReturns200WithOkMessage() throws Exception {
        when(response.getWriter()).thenReturn(printWriter);

        statusServlet.doGet(request, response);

        verify(response).setStatus(HttpServletResponse.SC_OK);
        verify(printWriter).print("OK DUDE");

    }

}
